Bungle TLS on x86-64.

+++ kalloc.h
@@ -6,7 +6,7 @@ struct kmem {
   char name[MAXNAME];
   struct spinlock lock;
   struct run *freelist;
-  uint nfree;
+  u64 nfree;
 } __attribute__ ((aligned (CACHELINE)));
 
 extern struct kmem kmems[NCPU];
